fix: widen lcd_orientation to uint16_t
Previously stored as uint8_t, which truncated 270 to 14 on assignment.
set_display_defaults() then saw 14 as invalid and reset the field to 0,
so coolerdash sent "orientation": 0 to CoolerControl on every refresh.

Only 270 is affected; 90 and 180 fit in uint8_t.

No other call sites touch the field; CoolerControl JSON layer auto-promotes.

+++ src/device/config.c
@@ -1072,7 +1072,7 @@ static void load_display_from_json(json_t *root, Config *config)
     {
         int val = (int)json_integer_value(orientation);
         if (is_valid_orientation(val))
-            config->lcd_orientation = (uint8_t)val;
+            config->lcd_orientation = (uint16_t)val;
     }
 
     json_t *width = json_object_get(display, "width");
